The chord.C file now contains rules L1, L2, L3, SU1, SU2, SR1, SR2, SR3, SR4.
We fiddled with rules SR? and F? (for fingers). The chord.txt files is now uptodate.

Created an aggregate element (from which an up-to-date aggregate can be pulled, and which blocks its pull output until it has a fresh aggregate to yield)

Created a delete element that looks up an entry by a key in one of its indices and then removes that entry from all indices.  The entry remains in the table (in its queue) until it expires.

Table now has a remove(index, key) method, as well as a COUNT aggregate function.

Tests have been written to play with successor updates, although Deletes are not thoroughly tested yet.