Merged in Petros' branch with bug fixes, revisions 1276:1380.
Just a review:

Fixed bug in TimedPullPush in which counter was incremented excessively, every time a run was rescheduled, regardless of success of the previous time.

Fixed the new planner to ensure it uses both old and new watch modifier specifications.

Added unit tests for the count of local periodics, as well as fast counter-like increments in tables.

Fixed double invocation of removeDerivatives in Table2.

Fixed initialization of Tuple::EMPTY to ensure it is allocated before use.

Added caveat in user guide about subscribing to table tuples.

Created a watchedTable.olg test file that shows whether insertEvents are caught for tables that are not read in the main program.

Added support for dummy rules watching table events.

Changed names in the namelessPeriodics.olg test file to make them more intuitive.